requirements.txt
openapi-typed
PyYAML
stone
